+++ src/common/Axis/RadialAxis/RadialAxisTickSeries/RadialAxisTickLabel.tsx
@@ -2,11 +2,13 @@ import type { FC, SVGTextElementAttributes } from 'react';
 import React, { useMemo } from 'react';
 
 import { formatValue } from '@/common/utils/formatting';
+import classNames from 'classnames';
+import css from './RadialAxisTickLabel.module.css';
 
 const rad2deg = (angle: number) => (angle * 180) / Math.PI;
 
 export interface RadialAxisTickLabelProps
-  extends Omit<SVGTextElementAttributes<SVGTextElement>, 'format'> {
+  extends Omit<SVGTextElementAttributes<SVGTextElement>, 'format' | 'onClick'> {
   /**
    * Data to render.
    */
@@ -66,6 +68,11 @@ export interface RadialAxisTickLabelProps
    * Format tooltip title on hover label.
    */
   formatTooltip?: (value: any, index: number) => any | string;
+
+  /**
+   * Click handler for the label.
+   */
+  onClick?: (event: React.MouseEvent<SVGGElement>, value: any) => void;
 }
 
 export const RadialAxisTickLabel: FC<Partial<RadialAxisTickLabelProps>> = ({
@@ -81,6 +88,7 @@ export const RadialAxisTickLabel: FC<Partial<RadialAxisTickLabelProps>> = ({
   lineSize,
   index,
   formatTooltip,
+  onClick,
   ...rest
 }) => {
   const { transform, textAnchor } = useMemo(() => {
@@ -125,7 +133,13 @@ export const RadialAxisTickLabel: FC<Partial<RadialAxisTickLabelProps>> = ({
     typeof formatTooltip === 'function' ? formatTooltip(data, index) : text;
 
   return (
-    <g transform={transform}>
+    <g
+      transform={transform}
+      className={classNames({
+        [css.clickable]: !!onClick
+      })}
+      onClick={(event) => onClick?.(event, data)}
+    >
       <title>{titleHover}</title>
       <text
         dy="0.35em"